Type cleanup + move RB_SetBlen to gl_utils.c
